# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nextstep Dashboard API
  version: 1.1.0
  description: >-
    Contract for the multi-tenant SML dashboard, production administration,
    distributed report execution, LINE delivery, and LIFF viewer. Mutation
    requests authenticated by cookie also require X-CSRF-Token.
servers:
  - url: https://dashboard.nextstep-soft.com
...
paths:
...
  /api/v1/admin/tenants/{tenantId}/recipients:
    parameters: [{$ref: '#/components/parameters/TenantID'}]
    get:
      tags: [Admin configuration]
      operationId: listRecipients
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200':
          description: Tenant recipients.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecipientPage'}
    post:
      tags: [Admin configuration]
      operationId: createRecipientInvitation
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecipientInput'}
      responses:
        '201':
          description: Pending recipient membership created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Recipient'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/tenants/{tenantId}/recipients/{recipientId}/permissions:
    put:
      tags: [Admin configuration]
      operationId: replaceRecipientPermissions
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PermissionUpdate'}
      responses:
        '200':
          description: Entire permission set replaced atomically.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Recipient'}
        '409': {$ref: '#/components/responses/Conflict'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/tenants/{tenantId}/recipients/{recipientId}/invitation:
    post:
      tags: [Admin configuration]
      operationId: reissueRecipientInvitation
      description: Replaces the unused invitation for a pending recipient. The previous invitation reference becomes invalid immediately.
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '200':
          description: Pending recipient with a newly issued one-time invitation URL.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Recipient'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/tenants/{tenantId}/recipients/{recipientId}/permission-dependencies:
    get:
      tags: [Admin configuration]
      operationId: getRecipientPermissionDependencies
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
      responses:
        '200':
          description: Active schedules that currently require each recipient permission. Schedule details are bounded to five per report.
          content: {application/json: {schema: {$ref: '#/components/schemas/PermissionDependencies'}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/admin/tenants/{tenantId}/schedule-recipient-options/query:
    post:
      tags: [Admin configuration]
      operationId: queryScheduleRecipientOptions
      description: Read-only, tenant-scoped recipient eligibility query. This operation never changes a schedule.
      security: [{AdminSession: []}]
      parameters: [{$ref: '#/components/parameters/TenantID'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ScheduleRecipientOptionsInput'}}}
      responses:
        '200':
          description: Paginated recipients plus every currently selected recipient, including invalid selections.
          content: {application/json: {schema: {$ref: '#/components/schemas/ScheduleRecipientOptions'}}}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/tenants/{tenantId}/recipients/{recipientId}:
    get:
      tags: [Admin configuration]
      operationId: getTenantRecipient
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
      responses:
        '200':
          description: Recipient scoped to the requested tenant membership.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Recipient'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Admin configuration]
      operationId: revokeTenantRecipient
      description: Revokes the recipient membership for this tenant while retaining audit and delivery history.
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {description: Recipient membership revoked.}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: nextstep_admin_session
    ViewerSession:
      type: apiKey
      in: cookie
      name: nextstep_viewer_session
  parameters:
    TenantID: {name: tenantId, in: path, required: true, schema: {type: string, format: uuid}}
    RecipientID: {name: recipientId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 200}}
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
  responses:
...
    NotFound: {$ref: '#/components/responses/Problem404'}
    Conflict: {$ref: '#/components/responses/Problem409'}
...
    ValidationFailed: {$ref: '#/components/responses/Problem422'}
...
    Problem404: {description: Resource not found., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
    Problem409: {description: 'Duplicate request, state conflict, or optimistic lock conflict.', content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
    Problem422: {description: Input is syntactically valid but violates a business rule., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
  schemas:
...
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            retryable: {type: boolean}
            fieldErrors: {type: array, items: {$ref: '#/components/schemas/FieldError'}}
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: {type: string}
        code: {type: string}
        message: {type: string}
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: string, nullable: true}
        hasMore: {type: boolean}
...
    RecipientInput:
      type: object
      additionalProperties: false
      required: [invitationLabel]
      properties: {invitationLabel: {type: string, minLength: 1, maxLength: 160}}
    Recipient:
      type: object
      required: [id, status, displayName, reportKeys, permissionsVersion, createdAt]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [PENDING, ACTIVE, REVOKED]}
        displayName: {type: string}
        reportKeys: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/ReportKey'}}
        permissionsVersion: {type: integer, minimum: 1}
        verifiedAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
        invitationUrl: {type: string, format: uri, nullable: true, description: Returned only when a pending invitation is created.}
    RecipientPage:
      type: object
      required: [data, page]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Recipient'}}
        page: {$ref: '#/components/schemas/PageInfo'}
    PermissionUpdate:
      type: object
      additionalProperties: false
      required: [reportKeys, version]
      properties:
        reportKeys: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/ReportKey'}}
        version: {type: integer, minimum: 1}
    PermissionDependencies:
      type: object
      additionalProperties: false
      required: [recipientId, permissionsVersion, items]
      properties:
        recipientId: {type: string, format: uuid}
        permissionsVersion: {type: integer, minimum: 1}
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [reportKey, activeScheduleCount, schedules, additionalCount]
            properties:
              reportKey: {$ref: '#/components/schemas/ReportKey'}
              activeScheduleCount: {type: integer, minimum: 1}
              schedules:
                type: array
                maxItems: 5
                items:
                  type: object
                  additionalProperties: false
                  required: [id, name]
                  properties:
                    id: {type: string, format: uuid}
                    name: {type: string}
              additionalCount: {type: integer, minimum: 0}
    ScheduleRecipientOptionsInput:
      type: object
      additionalProperties: false
      required: [reportKeys, selectedRecipientIds, search, page, pageSize]
      properties:
        reportKeys: {type: array, minItems: 1, maxItems: 10, uniqueItems: true, items: {$ref: '#/components/schemas/ReportKey'}}
        selectedRecipientIds: {type: array, maxItems: 500, uniqueItems: true, items: {type: string, format: uuid}}
        search: {type: string, maxLength: 160}
        page: {type: integer, minimum: 0}
        pageSize: {type: integer, minimum: 1, maximum: 100}
    ScheduleRecipientOption:
      allOf:
        - {$ref: '#/components/schemas/Recipient'}
        - type: object
          required: [eligible, missingReportKeys]
          properties:
            eligible: {type: boolean}
            missingReportKeys: {type: array, items: {$ref: '#/components/schemas/ReportKey'}}
    ScheduleRecipientOptions:
      type: object
      additionalProperties: false
      required: [data, selected, page, pageSize, total, hasMore]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/ScheduleRecipientOption'}}
        selected: {type: array, maxItems: 500, items: {$ref: '#/components/schemas/ScheduleRecipientOption'}}
        page: {type: integer, minimum: 0}
        pageSize: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
        hasMore: {type: boolean}
    ReportKey:
      type: string
      enum: [sales_goods_services, purchase_goods_payables, gross_profit_by_product, gross_profit_by_ar_customer, stock_balance, stock_reorder, ar_customer_movement, ar_debt_receipt, cash_bank_receipts, cash_bank_payments]
...